version 1.1.0
增加了签到信息判断
