--- stu_num/stu_num.csv
@@ -1,4 +1,0 @@
-id,name
-1,张三
-2,李四
-3,王五
